Cleanup moto/flask dependencies
