+++ pyproject.toml
@@ -80,13 +80,11 @@ safety = "^3.3.0"
 pylint = "^3.3.7"
 flake8-pyproject = "^1.2.3"
 httpx = "^0.27.2"
-moto = "^5.1.3"
-flask = "^3.0.3"
-flask-cors = "^5.0.1"
 sphinx = "^8.2.3"
 sphinx-rtd-theme = "^3.0.1"
 sqlalchemy-utils = "^0.41.2"
 click = "8.1.8"
+responses = "^0.25.7"
 
 [tool.poetry.requires-plugins]
 poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }
