+++ services/catalog/pyproject.toml
@@ -55,9 +55,9 @@ pytest-error-for-skips = "^2.0.2"
 pytest-docker = "^3.1.0"
 pytest-dotenv = "^0.5.2"
 pytest-mock = "^3.14.0"
-moto = "^5.1.0"
-flask = "^3.0.3"
-flask-cors = "^5.0.1"
+moto = "^5.1.4"
+flask = "^3.1.1" # for moto
+flask-cors = "^6.0.0" # for moto
 sqlalchemy-utils = "^0.41.2"
 psycopg2 = "^2.9.9"
 pytest-httpx = ">=0.32,<0.35"
