+++ services/common/pyproject.toml
@@ -81,7 +81,10 @@ pytest-docker = "^3.1.0"
 pytest-error-for-skips = "^2.0.2"
 pytest-httpx = "^0.34.0"
 pytest-mock = "^3.14.0"
-responses = "^0.25.3"
+responses = "^0.25.7"
+moto = "^5.1.4"
+flask = "^3.1.1" # for moto
+flask-cors = "^6.0.0" # for moto
 
 [tool.black]
 line-length = 120
