+++ services/common/tests/s3_storage_handler/helpers.py
@@ -1,0 +1,54 @@
+# Copyright 2025 CS Group
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Various helpers for tests and tests fixtures."""
+
+import os
+import os.path as osp
+from pathlib import Path
+
+import yaml
+
+RESOURCES_FOLDER = Path(osp.realpath(osp.dirname(__file__))) / ".." / "testresources"
+
+
+def export_aws_credentials():
+    """Export AWS credentials as environment variables for testing purposes.
+
+    This function sets the following environment variables with dummy values for AWS credentials:
+    - AWS_ACCESS_KEY_ID
+    - AWS_SECRET_ACCESS_KEY
+    - AWS_SECURITY_TOKEN
+    - AWS_SESSION_TOKEN
+    - AWS_DEFAULT_REGION
+
+    Note: This function is intended for testing purposes only, and it should not be used in production.
+
+    Returns:
+        None
+
+    Raises:
+        None
+    """
+    with open(RESOURCES_FOLDER / "s3" / "s3.yml", encoding="utf-8") as f:
+        s3_config = yaml.safe_load(f)
+        os.environ.update(s3_config["s3"])
+
+
+def clear_aws_credentials():
+    """Clear AWS credentials from environment variables."""
+    with open(RESOURCES_FOLDER / "s3" / "s3.yml", encoding="utf-8") as f:
+        s3_config = yaml.safe_load(f)
+        for env_var in list(s3_config["s3"].keys()):
+            del os.environ[env_var]
