+++ services/common/tests/testresources/s3/full_s3_storage_handler_test/subdir_1/subsubdir_2/subsubdir_2_file2
