+++ services/common/tests/testresources/s3/full_s3_storage_handler_test/subdir_2/subdir_2_file2
