+++ services/frontend/pyproject.toml
@@ -62,7 +62,7 @@ httpx = "^0.27.2"
 # Tools
 requests = "^2.32.3"
 # Test tools
-responses = "^0.25.3"
+responses = "^0.25.7"
 pylint = "^3.3.0"
